+++ CHANGELOG.md
@@ -24,6 +24,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `ui.diff.format`
   - `ui.diff.tool`
 
+* The deprecated `commit_id.normal_hex()` template method has been removed.
+
 ### Deprecations
 
 * The `git_head()` and `git_refs()` functions will be removed from revsets and